rename oldbench to bench1
